Split out typ pretty printer from pretty printer for module
